Fixing bug because paths already come back relative
This is/was causing a bug in symfony - many resources weren't being returned, so routes weren't being generated.

Specifically, since the paths were already relative, the substr returned false, and only one path would be returned per resource.

However, even though this fixes the problem - so that router:debug shows the asset in symfony - the route is still not actually found when you browse to the site or run your tests. I'm hoping this will help someone else recognize why that might be the case.