[CompassFilter] Fix empty ProcessBuilder arguments
When passing `--sass-dir` and `--css-dir` options to the compass binary,
the filter uses empty arguments.

But using empty arguments on windows seems to screw the way arguments
are passed to the binary (they are translated into "" by the process
builder which seems to be purely ignored by the shell), therefore the
following arguments are concatenated.

This leads compass to raise a "You must compile individual stylesheets
from the project directory." error.

Adding a '.' instead of just '' fixes this issue on Windows 7 Pro using
Compass 0.12.2 (Alnilam).