Fix broken compass test in ScssphpFilter
The scss lib throws an exception if you reference an undefined mixin,
which is what this test does. I've split it into two tests - one
checking that the mixin works when compass is enabled, and the other
checking that it throws an exception when the mixin is referenced and
compass is not enabled.