Update SeparatorFilter.php
Appending the separator to the end of the file will not work in cases such as a javascript file that ends in a line comment ( "//" ) with no trailing newline. The separator will be commented out. The last line of a file is a common place to put a comment containing a sourcemap for minified JS.

Rather than make assumptions about whether the user means for any separator to be preceded by a newline, or for the default separator to be '\n;', other libraries that do this kind of thing, such as jsqueeze, simply put the separator at the beginning of the file, where it cannot be preceded by a line comment.